+++ packages/plugin-audio/web/src/audio-component.tsx
@@ -7,15 +7,19 @@ import type {
   ComponentData,
   Helpers,
   RichContentTheme,
+  Store,
   TranslationFunction,
 } from 'wix-rich-content-common';
 import { MediaItemErrorMsg } from 'wix-rich-content-ui-components';
+import type { ContentBlock } from 'draft-js';
 
 interface Props {
   componentData: ComponentData;
   settings: AudioPluginEditorConfig;
   theme: RichContentTheme;
   helpers: Helpers;
+  store?: Store;
+  block?: ContentBlock;
   isMobile: boolean;
   t?: TranslationFunction;
   isLoading?: boolean;
@@ -26,7 +30,8 @@ class AudioComponent extends React.Component<Props> {
   static type = { AUDIO_TYPE };
 
   render() {
-    const { componentData, settings, theme, helpers, isMobile, isLoading, t } = this.props;
+    const { componentData, settings, theme, helpers, isMobile, isLoading, t, store, block } =
+      this.props;
     return (
       <>
         <AudioViewer
@@ -36,6 +41,8 @@ class AudioComponent extends React.Component<Props> {
           helpers={helpers}
           isMobile={isMobile}
           isLoading={componentData?.tempData || isLoading}
+          store={store}
+          block={block}
           disabled
         />
         {componentData?.error && <MediaItemErrorMsg error={componentData?.error} t={t} />}