+++ packages/plugin-audio/web/src/audio-viewer.tsx
@@ -1,18 +1,21 @@
 import type { FC } from 'react';
 import React, { useMemo } from 'react';
-import type { RichContentTheme, ComponentData, Helpers } from 'wix-rich-content-common';
+import type { RichContentTheme, ComponentData, Helpers, Store } from 'wix-rich-content-common';
 import { mergeStyles } from 'wix-rich-content-common';
 import type { AudioPluginViewerConfig } from './types';
 import ReactPlayerWrapper from './ReactPlayerWrapper';
 import styles from '../statics/styles/audio.rtlignore.scss';
 import AudioEmbedViewer from './AudioEmbedViewer';
 import { getAudioSrc } from './consts';
+import type { ContentBlock } from 'draft-js';
 
 interface Props {
   componentData: ComponentData;
   settings: AudioPluginViewerConfig;
   theme: RichContentTheme;
   helpers: Helpers;
+  store?: Store;
+  block?: ContentBlock;
   isMobile: boolean;
   disabled?: boolean;
   isLoading?: boolean;
@@ -26,11 +29,19 @@ const AudioViewer: FC<Props> = ({
   disabled,
   componentData,
   isLoading,
+  store,
+  block,
 }) => {
   const classes = useMemo(() => mergeStyles({ styles, theme }), [theme]);
   const { audio = {}, authorName, name, coverImage, disableDownload } = componentData;
   const url = getAudioSrc(audio?.src, settings);
   const shouldEmbed = componentData.html;
+  const saveDurationToData = duration =>
+    store?.update(
+      'componentData',
+      { ...componentData, audio: { ...componentData.audio, duration } },
+      block?.getKey()
+    );
 
   return (
     <div className={classes.container} data-hook="audioViewer">
@@ -47,6 +58,7 @@ const AudioViewer: FC<Props> = ({
           showControls={!!audio?.src?.url}
           disabled={disabled}
           isLoading={isLoading}
+          saveDurationToData={saveDurationToData}
         />
       ) : (
         <AudioEmbedViewer
@@ -54,6 +66,7 @@ const AudioViewer: FC<Props> = ({
           settings={settings}
           theme={theme}
           isMobile={isMobile}
+          helpers={helpers}
         />
       )}
     </div>