+++ CHANGELOG.md
@@ -17,6 +17,13 @@
     Click to see more.
   </summary>
 
+
+
+</details>
+  <hr/>
+
+## 8.70.41 (April 17, 2022)
+
 ### :bug: Bug Fix
 - `video`
   - [#3737](https://github.com/wix/ricos/pull/#3737) fix "getInternalPlayer().play(), play is not a function"
@@ -24,9 +31,6 @@
   - [#3738](https://github.com/wix/ricos/pull/3738) fix audio tags issue & added BI
 - `ui-components`
   - [#3745](https://github.com/wix/ricos/pull/#3745) add loader to SettingsAddItem component
-
-</details>
-  <hr/>
 
 ## 8.70.40 (April 13, 2022)